doa/python-doa-lib/pyproject.toml
[tool.poetry]
authors = ["Shin Saito <shinsa82@gmail.com>"]
description = "Tackle-DiVA-DOA: Analyzer Libraray for Assessment and Migration to PostgreSQL"
license = "Apache-2.0"
name = "doa"
version = "2.4.0"

[tool.poetry.dependencies]
PyYAML = "^6.0"
antlr4-python3-runtime = "^4.10"
more-itertools = "^8.12.0"
psycopg2 = "^2.9.3"
python = "^3.10"
rich = "^12.2.0"
typer = "^0.4.1"
wasabi = "^0.9.1"

[tool.poetry.dev-dependencies]
pytest = "*"

[tool.doa]
image-name = "diva-doa"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
